Fixed some bugs, Favorite list still not working

+++ build.sh
@@ -11,25 +11,36 @@ rm -Rf bitacorapp-win32-x64/
 echo "Compiling..."
 electron-packager ./ bitacorapp --platform=$PLATFORM --arch=$ARCH --version=$ELVERSION --icon bitmap.ico --overwrite
 echo "Entering new directory"
-cd bitacorapp-win32-x64
-echo "Moving required programs..."
-mv resources/app/bin ./bin
-cd bin
-rm images/*
-rm posts/*
-site clean
-cd ..
-mv resources/app/bndb.bat .
-mv resources/app/bndb.exe .
-mv resources/app/killer.exe .
-mv resources/app/kill.bat .
-cd ..
-cp resources/database_empty bitacorapp-win32-x64/resources/database
-cp resources/default_note   bitacorapp-win32-x64/resources/default_note
-cp bitacorapp-win32-x64 ../
-cd ..
-cd bitacorapp-win32-x64
-echo "Executing program..."
-./bitacorapp.exe
+
+if [ -d bitacorapp-win32-x64 ]; then
+    cd bitacorapp-win32-x64
+    echo "Moving required programs..."
+    mv resources/app/bin ./bin
+    mv resources/app/bndb.bat .
+    mv resources/app/kill.bat .
+    cd bin
+    echo "Removing previous images..."
+    rm images/*
+    cd ../.. # exits bitacorapp
+    cp resources/default_note   bitacorapp-win32-x64/resources/default_note
+    echo "Compiling backend..."
+    cd ../bndb # exits beenotes, enters bndb
+    stack build
+    cp .stack-work/install/i386-windows/lts-3.7/7.10.2/bin/bndb.exe ../beenotes/bitacorapp-win32-x64/bndb.exe
+    echo "Compiling backend killer..."
+    cd ../killer # exits bndb
+    cp .stack-work/install/i386-windows/lts-3.15/7.10.2/bin/killer.exe ../beenotes/bitacorapp-win32-x64/killer.exe
+    echo "Copying screenshot program..."
+    cd ../BNScreenShot
+    cp BNScreenShot/bin/Debug/BNScreenShot.exe ../beenotes/bitacorapp-win32-x64/bin/bnsc.exe
+    echo "Copying syntax highlighter..."
+    cd ../highlighting-kate-0.6
+    cp .stack-work/install/i386-windows/lts-3.16/7.10.2/bin/highlighting-kate.exe ../beenotes/bitacorapp-win32-x64/bin/highlighting-kate.exe
+    cd ../beenotes/bitacorapp-win32-x64
+    echo "Clean up..."
+    rm resources/app/*sh
+    echo "Executing program..."
+    ./bitacorapp.exe
+fi
